# doc/openapi.yaml
openapi: 3.0.3
info:
  title: CoZaDzban API
  description: |
    ## Overview
    CoZaDzban is a link-collecting application. Users submit a URL and a language
    tag; the server fetches the target page and automatically enriches the record
    with title, thumbnail, favicon, lead, and best-effort publication date before persisting it.

    ## Components
    1. **REST API** — CRUD endpoints for `Article` resources (this spec).
    2. **PostgreSQL database** — all data persisted via Spring Data JDBC.
    3. **Flyway migrations** — all schema changes are managed as versioned SQL
       migration scripts located in `src/main/resources/db/migration/`.
       No DDL is generated or executed outside of Flyway.
    4. **RSS feed** — `GET /rss` exposes articles as an RSS 2.0 feed;
       filterable per language with `?lang=<BCP-47>`.
    5. **Vaadin UI** — server-side lazy feed of articles (separate from
       this REST spec; rendered at the root path `/`).

    6. **Frontend authentication** — Vaadin UI starts OAuth Authorization Code + PKCE
       via backend redirect (`GET /auth/login`) and uses server-side session cookie.
    7. **Admin pages** — Vaadin UI at `/admin` for ADMIN users to manage allowlist
       (add/soft-delete/restore/update users) and at `/admin/article-content-cache`
       to inspect, refresh, edit, and purge preserved article content cache entries,
       using server-side service calls.
...
       imports remain non-interactive.
...
    ## Implementation notes
...

...
servers:
...
paths:
  /api/articles:

    get:
      tags: [Articles]
      summary: List articles (paginated) or check whether a URL exists
      operationId: listArticles
      security: []
      parameters:
        - name: existsUrl
          in: query
          description: When present, returns only whether the canonicalized URL already exists.
          required: false
          schema:
            type: string
            format: uri
            maxLength: 2048
            pattern: '^https?://.+'
        - name: page
          in: query
          description: Zero-based page index
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          description: Number of articles per page
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: sort
          in: query
          description: >
            Sort expression: field,direction — e.g. createdAt,desc.
            Permitted fields: id, createdAt, publishedAt, language, title, url.
            Unknown field returns 400 Bad Request.
          schema:
            type: string
            enum:
              - id,asc
              - id,desc
              - createdAt,asc
              - createdAt,desc
              - publishedAt,asc
              - publishedAt,desc
              - language,asc
              - language,desc
              - title,asc
              - title,desc
              - url,asc
              - url,desc
            example: createdAt,desc
        - name: language
          in: query
          description: Normalized BCP-47-like language filter, e.g. pl or en-gb
          required: false
          schema:
            type: string
            maxLength: 20
            pattern: '^[a-z]{2,3}(-[a-z0-9]{2,8})*$'
            example: pl
        - name: publishedFrom
          in: query
          description: Inclusive lower bound for publishedAt
          required: false
          schema:
            type: string
            format: date-time
        - name: publishedTo
          in: query
          description: Inclusive upper bound for publishedAt
          required: false
          schema:
            type: string
            format: date-time
        - name: createdFrom
          in: query
          description: Inclusive lower bound for createdAt
          required: false
          schema:
            type: string
            format: date-time
        - name: createdTo
          in: query
          description: Inclusive upper bound for createdAt
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Paginated list of articles
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ArticlePage'
                  - $ref: '#/components/schemas/ArticleUrlExistsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

    post:
      tags: [Articles]
      summary: Create a new article
      operationId: createArticle
      security:
        - googleOidc: []
        - machineApiKey: []
      description: >
        Accepts a URL and a BCP-47 language tag.
        The server fetches the page at the given URL and automatically
        populates title, thumbnail, favicon, lead, and best-effort publishedAt before
        persisting. Creator is resolved from the authenticated DB user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleInput'
            example:
              url: https://example.com/some-article
              language: pl
      responses:
        '201':
          description: Article created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /auth/me:

    get:
      tags: [Authentication]
      summary: Get current authenticated user
      operationId: getCurrentUser
      security:
        - cookieAuth: []
      description: >
        Returns email, nullable role, and allowlist state of the currently
        authenticated user from server session. Authenticated users who are not
        active in the allowlist receive `allowlisted=false` and `role=null`.
      responses:
        '200':
          description: Current user details
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                    example: user@example.com
                  role:
                    type: string
                    nullable: true
                    enum: [USER, ADMIN]
                    example: ADMIN
                  allowlisted:
                    type: boolean
                    example: true
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:

    post:
      tags: [Authentication]
      summary: Clear authentication session
      operationId: logout
      security:
        - cookieAuth: []
      description: >
        Invalidates server session and clears session cookie.
        Returns 204 No Content.
      responses:
        '204':
          description: Logged out successfully; cookie cleared
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/articles/{id}:

    parameters:
      - $ref: '#/components/parameters/ArticleId'

    get:
      tags: [Articles]
      summary: Get article by ID
      operationId: getArticle
      security: []
      responses:
        '200':
          description: Article found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Articles]
      summary: Replace article input fields
      operationId: replaceArticle
      security:
        - googleOidc: []
      description: >
        Replaces url, language and optional quote. The server re-fetches the
        target page to refresh title, thumbnail, lead, and best-effort
        publishedAt unless publishedAt is explicitly supplied.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleInput'
      responses:
        '200':
          description: Article replaced
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    patch:
      tags: [Articles]
      summary: Partially update an article
      operationId: patchArticle
      security:
        - googleOidc: []
        - machineApiKey: []
      description: >
        Updates only the provided fields. Updatable fields: url, language, quote, publishedAt,
        and internal preserved content.
        Setting quote to null explicitly clears it; omitting quote leaves it unchanged.
        Setting publishedAt to null explicitly clears it; omitting publishedAt leaves it unchanged unless url changes and no override is supplied.
        Setting content stores/replaces the preserved content cache; setting content to null or blank clears it.
        If url is changed the server re-fetches the page to refresh the enriched
        fields. Fully atomic — if enrichment fails the server responds with 422
        and the existing record is left unchanged.
        Null for url or language is invalid and returns 400.
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/ArticlePatch'
      responses:
        '200':
          description: Article updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      tags: [Articles]
      summary: Delete an article
      operationId: deleteArticle
      security:
        - googleOidc: []
      responses:
        '204':
          description: Article deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    googleOidc:
      type: openIdConnect
      openIdConnectUrl: https://accounts.google.com/.well-known/openid-configuration

    machineApiKey:
      type: apiKey
      in: header
      name: X-CoZaDzban-M2M-Key
      description: Machine credential used by the remote Facebook importer to write articles

    cookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Server-managed authenticated session cookie
  parameters:
    ArticleId:
      name: id
      in: path
      required: true
      description: Unique article identifier
      schema:
        type: integer
        format: int64
        example: 42
  schemas:
    ArticleInput:
      type: object
      required: [url, language]
      properties:
        url:
          type: string
          format: uri
          minLength: 1
          maxLength: 2048
          pattern: '^https?://.+'
          description: Full URL of the article page to fetch and store (http/https only). The server persists the canonicalized URL and removes tracking-only query parameters such as fbclid, utm_*, gclid, dclid, and msclkid before uniqueness checks.
          example: https://example.com/some-article
        language:
          type: string
          minLength: 1
          maxLength: 20
          pattern: '^[a-z]{2,3}(-[a-z0-9]{2,8})*$'
          description: BCP-47 language tag identifying the article's language
          example: pl
        quote:
          type: string
          nullable: true
          minLength: 1
          maxLength: 2000
          description: Optional user-provided quote or comment about the article
          example: "This perfectly captures the absurdity."
        publishedAt:
          type: string
          format: date-time
          nullable: true
          description: Optional manual source publication timestamp. If omitted on create, enrichment attempts to populate it.
          example: "2026-05-02T08:30:00Z"
    ArticleUrlExistsResponse:
      type: object
      required: [exists]
      properties:
        exists:
          type: boolean
          example: true
    ArticlePatch:
      type: object
      description: JSON Merge Patch document; omitted fields are left unchanged
      properties:
        url:
          type: string
          format: uri
          minLength: 1
          maxLength: 2048
          pattern: '^https?://.+'
          example: https://example.com/updated-article
        language:
          type: string
          minLength: 1
          maxLength: 20
          pattern: '^[a-z]{2,3}(-[a-z0-9]{2,8})*$'
          description: BCP-47 language tag
          example: en
        quote:
          type: string
          nullable: true
          minLength: 1
          maxLength: 2000
          description: User-provided quote; null explicitly clears it; empty string is invalid
          example: "This perfectly captures the absurdity."
        publishedAt:
          type: string
          format: date-time
          nullable: true
          description: Source publication timestamp; null explicitly clears it
          example: "2026-05-02T08:30:00Z"
        content:
          type: string
          nullable: true
          description: Internal preserved content cache value; null or blank clears cached content. Not exposed in article responses or RSS.
    Article:
      type: object
      description: >
        Article response. `createdBy` is included only for authenticated requests
        and omitted for anonymous/public responses and RSS.
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          example: 42
        url:
          type: string
          format: uri
          example: https://example.com/some-article
        language:
          type: string
          description: BCP-47 language tag
          example: pl
        title:
          type: string
          description: Title fetched from the article page
          nullable: true
          example: "An Interesting Article"
        thumbnail:
          type: string
          format: uri
          description: Thumbnail URL fetched from the article page (og:image or similar)
          nullable: true
          example: https://example.com/images/thumb.jpg
        favicon:
          type: string
          description: Local cached favicon URL downloaded from source site icon metadata
          nullable: true
          example: /favicons/2f1a0c8d9e7b6a5c4d3e2f10987654321abcdeffedcba9876543210012345678.png
        lead:
          type: string
          description: Lead/excerpt fetched from the article page (og:description or similar)
          nullable: true
          example: "A short summary of the article content."
        quote:
          type: string
          description: User-provided quote or comment about the article
          nullable: true
          example: "This perfectly captures the absurdity."
        aiSummary:
          type: string
          description: AI-generated summary — future feature, null until implemented
          nullable: true
        publishedAt:
          type: string
          format: date-time
          nullable: true
          description: Source article publication timestamp, fetched from metadata or edited manually
          example: "2026-05-02T08:30:00Z"
        createdAt:
          type: string
          format: date-time
          readOnly: true
          example: "2026-05-02T10:00:00Z"
        createdBy:
          $ref: '#/components/schemas/ArticleCreator'
    ArticleCreator:
      type: object
      readOnly: true
      description: Present only for authenticated article responses; never exposed in anonymous responses or RSS.
      properties:
        id:
          type: integer
          format: int64
          example: 1
        email:
          type: string
          format: email
          example: user@example.com
...
    ArticlePage:
      type: object
      description: Frozen page response shape for Article
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Article'
        pageable:
          $ref: '#/components/schemas/Pageable'
        totalElements:
          type: integer
          format: int64
          example: 150
        totalPages:
          type: integer
          example: 8
        size:
          type: integer
          example: 20
        number:
          type: integer
          description: Current zero-based page number
          example: 0
        numberOfElements:
          type: integer
          example: 20
        sort:
          $ref: '#/components/schemas/Sort'
        first:
          type: boolean
          example: true
        last:
          type: boolean
          example: false
        empty:
          type: boolean
          example: false
    Pageable:
      type: object
      properties:
        pageNumber:
          type: integer
          example: 0
        pageSize:
          type: integer
          example: 20
        offset:
          type: integer
          format: int64
          example: 0
        paged:
          type: boolean
          example: true
        unpaged:
          type: boolean
          example: false
        sort:
          $ref: '#/components/schemas/Sort'
    Sort:
      type: object
      properties:
        sorted:
          type: boolean
          example: true
        unsorted:
          type: boolean
          example: false
        empty:
          type: boolean
          example: false
    ProblemDetail:
      type: object
      description: RFC 9457 / Spring ProblemDetail error response
      required: [type, title, status, detail]
      additionalProperties: true
      properties:
        type:
          type: string
          format: uri
          example: https://example.com/problems/validation-error
        title:
          type: string
          example: Bad Request
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: "Field 'url' must be a valid URI"
        instance:
          type: string
          format: uri
          example: /api/articles
        conflictField:
          type: string
          description: Optional field name that caused a conflict
          example: url
        existingId:
          type: integer
          format: int64
          description: Optional existing resource identifier for conflict cases
          example: 42
        remainingAdmins:
          type: integer
          description: Optional count of admins that would remain after operation
          example: 0
  responses:
    BadRequest:
      description: Request payload is invalid or malformed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Conflict:
      description: Conflict with business rules
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          examples:
            articleUrlConflict:
              summary: Duplicate article URL
              value:
                type: https://cozadzban.pl/problems/article-url-conflict
                title: Article URL Already Exists
                status: 409
                detail: "Article with this URL already exists (canonicalized: <url>)."
                instance: /api/articles
                conflictField: url
                existingId: 42
            allowlistEmailConflict:
              summary: Duplicate allowlist email
              value:
                type: https://cozadzban.pl/problems/allowlist-email-conflict
                title: Allowlisted Email Already Exists
                status: 409
                detail: User with this email is already allowlisted.
                instance: /api/users
                conflictField: email
                existingId: 7
            lastAdminRequired:
              summary: Attempt to remove final admin
              value:
                type: https://cozadzban.pl/problems/last-admin-required
                title: Last Admin Required
                status: 409
                detail: Operation would leave the system without an ADMIN user.
                instance: /api/users/1
                remainingAdmins: 0
            facebookImportBusy:
              summary: Facebook import already running
              value:
                type: https://cozadzban.pl/problems/facebook-import-busy
                title: Facebook Import Already Running
                status: 409
                detail: Facebook import is already running.
                instance: /api/admin/facebook-import/run
            facebookImportNotRunning:
              summary: No Facebook import to terminate
              value:
                type: https://cozadzban.pl/problems/facebook-import-not-running
                title: No Active Facebook Import
                status: 409
                detail: No Facebook import job is currently running.
                instance: /api/admin/facebook-import/terminate
    Unauthorized:
      description: Missing or invalid authentication credentials
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Forbidden:
      description: Authenticated but not allowed to perform this operation
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    NotFound:
      description: The requested resource does not exist
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    UnprocessableEntity:
      description: The URL could not be fetched or its content could not be parsed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'